refactor(header): histories 헤더를 메인 헤더로 변경
